+++ src/components/Lane/LaneFooter.js
@@ -1,10 +1,10 @@
 import React from 'react'
 
-import {LaneFooter} from 'styles/Base'
+import {LaneFooter} from 'rt/styles/Base'
 
 import {
   CollapseBtn,
   ExpandBtn,
-} from 'styles/Elements'
+} from 'rt/styles/Elements'
 
 export default ({onClick, collapsed}) => <LaneFooter onClick={onClick}>{collapsed ? <ExpandBtn /> : <CollapseBtn />}</LaneFooter>