+++ src/widgets/index.js
@@ -1,0 +1,9 @@
+import DeleteButton from './DeleteButton'
+import EditableLabel from './EditableLabel'
+import InlineInput from './InlineInput'
+
+export default {
+  DeleteButton,
+  EditableLabel,
+  InlineInput
+}