+++ stories/EditableBoard.story.js
@@ -7,6 +7,10 @@ import Board from '../src'
 const data = require('./data/base.json')
 const smallData = require('./data/data-sort')
 
+const disallowAddingCardData =  {...data}
+disallowAddingCardData.lanes[0].title = 'Disallowed adding card'
+disallowAddingCardData.lanes[0].disallowAddingCard = true
+
 storiesOf('Editable Board', module)
   .add(
     'Add/Delete Cards',
@@ -53,4 +57,32 @@ storiesOf('Editable Board', module)
       )
     },
     {info: 'Allow adding new lane'}
+  )
+  .add(
+    'Disallow Adding Card for specific Lane',
+    () => {
+      return (
+        <Board
+          data={disallowAddingCardData}
+          editable
+        />
+      )
+    },
+    {info: 'Can hide the add card button on specific lanes'}
+  )
+  .add(
+    'Inline Edit Lane Title',
+    () => {
+      return (
+        <Board
+          data={smallData}
+          editable
+          canAddLanes
+          editLaneTitle
+          onLaneUpdate={ (laneId, data) => debug(`onLaneUpdate: ${laneId} -> ${data.title}`)}
+          onLaneAdd={t => debug('You added a line with title ' + t.title)}
+        />
+      )
+    },
+    {info: 'Allow edit lane title'}
   )